fix: resolve i18n/api conflicts and clean up route private imports
