+++ apps/web/src/lib/utils.ts
@@ -73,3 +73,16 @@ export const shuffleArray = <T>(originalArray: readonly T[]): T[] => {
 
   return array;
 };
+
+export const downloadFile = (blob: Blob, fileName: string) => {
+  const url = URL.createObjectURL(blob);
+
+  const link = document.createElement("a");
+  link.href = url;
+  link.download = fileName;
+
+  document.body.append(link);
+  link.click();
+  link.remove();
+  URL.revokeObjectURL(url);
+};
