+++ apps/web/src/routes/_protected.workspaces/$workspaceId/-components/row-actions.tsx
@@ -84,7 +84,7 @@ import {
   getPdfDownloadFileName,
 } from "@/routes/_protected.workspaces/$workspaceId/-components/row-actions.logic";
 import type { TableTreeNode } from "@/routes/_protected.workspaces/$workspaceId/-components/table/types";
-import { downloadFile } from "@/routes/_protected.workspaces/$workspaceId/-components/utils";
+import { downloadFile } from "@/lib/utils";
 import { useEntitiesCountLimit } from "@/routes/_protected.workspaces/$workspaceId/-hooks/use-limits";
 import { useRetryCell } from "@/routes/_protected.workspaces/$workspaceId/-hooks/use-retry-cell";
 import { useUploadVersion } from "@/routes/_protected.workspaces/$workspaceId/-hooks/use-upload-version";
