+++ apps/web/src/routes/_protected.workspaces/$workspaceId/-components/utils.ts
@@ -94,16 +94,3 @@ export const resolveOptionColor = (color: OptionColor): ColorVariants => {
 
 /** The 16 named preset color keys. */
 export const optionColors: readonly OptionColor[] = NAMED_COLORS;
-
-export const downloadFile = (blob: Blob, fileName: string) => {
-  const url = URL.createObjectURL(blob);
-
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = fileName;
-
-  document.body.append(link);
-  link.click();
-  link.remove();
-  URL.revokeObjectURL(url);
-};
