+++ apps/web/src/routes/_protected.workspaces/-components/pdf-viewer-controls.tsx
@@ -30,7 +30,7 @@ import {
   printPdfBuffer,
 } from "@/routes/_protected.workspaces/$workspaceId/-components/peek/peek-pdf-print";
 import { PeekPdfControls } from "@/routes/_protected.workspaces/$workspaceId/-components/peek/peek-pdf-viewer";
-import { downloadFile } from "@/routes/_protected.workspaces/$workspaceId/-components/utils";
+import { downloadFile } from "@/lib/utils";
 import { useWorkspaceStore } from "@/routes/_protected.workspaces/$workspaceId/-store";
 
 const SCALE_OFFSET_STEP = 0.2;
